+++ Optimizer/Options.cs
@@ -310,7 +310,8 @@ namespace Optimizer
                 if (CurrentOptions.LanguageCode == LanguageCode.KU) TranslationList = JObject.Parse(Properties.Resources.KU);
                 if (CurrentOptions.LanguageCode == LanguageCode.HU) TranslationList = JObject.Parse(Properties.Resources.HU);
                 if (CurrentOptions.LanguageCode == LanguageCode.RO) TranslationList = JObject.Parse(Properties.Resources.RO);
-            }
+                if (CurrentOptions.LanguageCode == LanguageCode.NL) TranslationList = JObject.Parse(Properties.Resources.NL);
+;            }
             catch (Exception ex)
             {
                 ErrorLogger.LogError("Options.LoadTranslation", ex.Message, ex.StackTrace);