+++ Optimizer/Program.cs
@@ -12,8 +12,8 @@ namespace Optimizer
         /* VERSION PROPERTIES */
         /* DO NOT LEAVE THEM EMPTY */
 
-        internal readonly static float Major = 13;
-        internal readonly static float Minor = 9;
+        internal readonly static float Major = 14;
+        internal readonly static float Minor = 0;
 
         internal readonly static bool EXPERIMENTAL_BUILD = false;
 
@@ -187,7 +187,7 @@ namespace Optimizer
                                 // resets configuration
                                 if (arg == "/reset")
                                 {
-                                    Utilities.ResetConfiguration(true);
+                                    Utilities.Repair(true);
                                     return;
                                 }